backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
 
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
passlib==1.7.4
bcrypt==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# RAG and ML dependencies
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2
scikit-learn==1.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0